modularized card server functions from routes
